NEW FEATURE:  AutoSync status now shown during gameplay with icon NEW FEATURE:  added autogen icon to select music NEW FEATURE:  toggle for easter eggs in Machine Options NEW FEATURE:  marvelous step timing togglable in Machine Options NEW FEATURE:  Grade AAAA (all marvelous) CHANGE:  Only show ScreenHowToPlay if at least one player chose easy

+++ stepmania/src/ScreenSelectMode.cpp
@@ -300,7 +300,7 @@ void ScreenSelectMode::AfterChange()
 	{
 		for( unsigned i=0; i<m_apPossibleModeChoices.size(); i++ )
 		{
-			if(i == m_ScrollingList.GetSelection())
+			if(i == (unsigned)m_ScrollingList.GetSelection())
 			{
 				m_Infotext[i].SetDiffuse(RageColor(1,1,1,1));
 			}
